// api/src/types/interface.d.ts
export interface SearchParams{
    keyword?: string;
    page?: number;
    limit?: number;
} 
export interface Student{
    studentCode?: string;
    studentNameKh?: string;
    studentNameEn?: string;
    dob?: Date;
    gender?: 'Male' | 'Female';
    email?: string;
    phoneNumber?: string;
    password?: string;
    changedByUser?: string;
    clientIp?: string;
    sessionInfo?: string;
    status?: 'Active' | 'Inactive';
}